gst-plugins-good: Autospec creation for update from version 1.20.1 to version 1.20.2
The latest bug-fix release in the 1.20 series is 1.20.2 and was released
on 2 May 2022.

See https://gstreamer.freedesktop.org/releases/1.20/ for the latest
version of this document.

Last updated: Monday 2 May 2022, 0:30 UTC (log)

Introduction

The GStreamer team is proud to announce a new major feature release in
the stable 1.x API series of your favourite cross-platform multimedia
framework!

As always, this release is again packed with many new features, bug

(NEWS truncated at 15 lines)